requirements.txt
geopandas==0.11.0
matplotlib==3.5.2
geomet==0.3.0
geojson==2.5.0
pandas==1.4.3
numpy==1.23.0
Pillow==9.2.0
scikit-learn==1.3.2
contextily==1.2.0
breizhcrops==0.0.4.1
tqdm==4.57.0
radiant-mlhub==0.4.1
dask==2022.6.1
Py6S==1.9.2
pyshp==2.3.0
openpyxl==3.0.10

jupyter==1.0.0
xarray==2022.3.0
salem==0.3.7
seaborn==0.11.2
Shapely==1.8.2
setuptools==56.0.0
rasterio==1.3.9
sentinelhub==3.6.3
lightly==1.4.11
tsaug==0.2.1
opencv-python==4.6.0.66
torchsummary==1.5.1
pytest==7.4.3
pytorch-lightning==1.9.5
tensorboard==2.15.1